requirements.txt
langchain<1.0.0
langchain-community<1.0.0
langchain-huggingface<1.0.0
chromadb
pypdf
sentence-transformers
streamlit
python-dotenv
pandas
numpy
langchain-chroma<1.0.0
langchain-google-genai<3.0.0